+++ server/src/definition.ts
@@ -1,0 +1,8 @@
+import { Node } from 'wollok-ts'
+
+export const getNodeDefinition = (node: Node): Node => {
+  if (node.kind === 'Reference') {
+    return node.target()!
+  }
+  return node
+}
